+++ packages/core/src/files/FilesLoaders.js
@@ -13,16 +13,19 @@ const path = require("path");
 const globule = require("globule");
 const watch = require("node-watch");
 const fsExtra = require("fs-extra");
-const { map, debounce, flatten } = require("lodash");
+const { map, debounce } = require("lodash");
+const isPromise = require("is-promise");
+
+const CollectionsLoader = require("./loaders/Collections");
+const RoutesLoader = require("./loaders/Routes");
+const Loader = require("./Loader");
 
 const {
-  collectionsFileToUse,
   babelRegisterDefaultOptions,
   getFilesGlobule,
-  validateFileContent,
+  isYamlFile,
+  readYamlFile,
 } = require("./helpers");
-
-const ROUTES_FOLDER = "routes";
 
 const OPTIONS = [
   {
@@ -68,14 +71,16 @@ class FilesLoaders {
   }
 
   constructor({ config, loadCollections, logger, loadRoutes, alerts }, extraOptions = {}) {
+    this._loaders = {};
     this._logger = logger;
+    this._alerts = alerts;
     this._loadCollections = loadCollections;
     this._loadRoutes = loadRoutes;
-    this._alerts = alerts;
-    this._deprecationAlerts = alerts.collection("deprecated");
-    this._collectionsAlerts = alerts.collection("collections");
-    this._routesAlerts = alerts.collection("routes");
-    this._routesFilesAlerts = this._routesAlerts.collection("file");
+
+    this._loggerLoaders = this._logger.namespace("loader");
+    this._alertsLoaders = alerts.collection("loader");
+    this._alertsLoad = alerts.collection("load");
+
     this._customRequireCache = extraOptions.requireCache;
     this._require = extraOptions.require || require;
     this._config = config;
@@ -86,18 +91,35 @@ class FilesLoaders {
       .addOptions(BABEL_REGISTER_OPTIONS);
     this._pathOption.onChange(this._onChangePathOption.bind(this));
     this._watchOption.onChange(this._onChangeWatchOption.bind(this));
+
+    this._loadLoader = this._loadLoader.bind(this);
+    this.createLoader = this.createLoader.bind(this);
   }
 
   init() {
+    this._collectionsLoader = new CollectionsLoader({
+      loadCollections: this._loadCollections,
+      createLoader: this.createLoader,
+      getBasePath: this._getPath.bind(this),
+    });
+    this._routesLoader = new RoutesLoader({
+      loadRoutes: this._loadRoutes,
+      createLoader: this.createLoader,
+      getBasePath: this._getPath.bind(this),
+    });
     this._enabled = this._enabledOption.value;
-    try {
-      if (this._enabled) {
-        this._loadFiles();
+    if (this._enabled) {
+      try {
+        return this._loadFiles();
+      } catch (error) {
+        return Promise.reject(error);
       }
-      return Promise.resolve();
-    } catch (err) {
-      return Promise.reject(err);
     }
+    return Promise.resolve();
+  }
+
+  reload() {
+    return this._loadFiles();
   }
 
   start() {
@@ -114,8 +136,19 @@ class FilesLoaders {
   }
 
   _readFile(filePath) {
-    const content = this._require(filePath);
-    return (content && content.default) || content;
+    if (isYamlFile(filePath)) {
+      this._logger.debug(`Reading yaml file ${filePath}`);
+      return readYamlFile(filePath);
+    }
+    this._logger.debug(`Reading file ${filePath}`);
+    return new Promise((resolve, reject) => {
+      try {
+        const content = this._require(filePath);
+        resolve((content && content.default) || content);
+      } catch (error) {
+        reject(error);
+      }
+    });
   }
 
   _cleanRequireCacheFolder() {
@@ -170,88 +203,59 @@ class FilesLoaders {
       );
     }
     this._cleanRequireCacheFolder();
-    this._loadRoutesFiles();
-    this._loadCollectionsFile();
+    return this._loadLoaders();
   }
 
-  _loadRoutesFiles() {
-    const routesPath = path.resolve(this._path, ROUTES_FOLDER);
-    try {
-      const routeFiles = globule.find({
-        src: getFilesGlobule(
-          this._babelRegisterOption.value,
-          this._babelRegisterOptionsOption.value
-        ),
-        srcBase: routesPath,
-        prefixBase: true,
-      });
-      this._routesAlerts.clean();
-      const routes = flatten(
-        routeFiles
-          .map((filePath) => {
-            const fileContent = this._readFile(filePath);
-            const fileErrors = validateFileContent(fileContent);
-            if (!!fileErrors) {
-              this._routesFilesAlerts.set(
-                filePath,
-                `Error loading routes from file ${filePath}: ${fileErrors}`
-              );
-              return null;
-            }
-            return fileContent;
+  _loadLoaders() {
+    this._alertsLoad.clean();
+    return Promise.all(map(this._loaders, this._loadLoader));
+  }
+
+  _loadLoader(loader) {
+    const filesToLoad = globule.find({
+      src: getFilesGlobule(
+        loader.src,
+        this._babelRegisterOption.value,
+        this._babelRegisterOptionsOption.value
+      ),
+      srcBase: this._getPath(),
+      prefixBase: true,
+    });
+
+    this._logger.silly(`Files to load for loader '${loader.id}': ${JSON.stringify(filesToLoad)}`);
+
+    return Promise.all(
+      filesToLoad.map((filePath) => {
+        return this._readFile(filePath)
+          .then((fileContent) => {
+            return {
+              path: filePath,
+              content: fileContent,
+            };
           })
-          .filter((fileContent) => !!fileContent)
-      );
-      this._loadRoutes(routes);
-      this._logger.silly(`Loaded routes from folder ${routesPath}`);
-    } catch (error) {
-      this._loadRoutes([]);
-      this._routesAlerts.set("error", `Error loading routes from folder ${routesPath}`, error);
-    }
-  }
-
-  _loadCollectionsFile() {
-    let collectionsFile = collectionsFileToUse(
-      this._path,
-      this._babelRegisterOption.value,
-      this._babelRegisterOptionsOption.value
-    );
-
-    if (collectionsFile) {
-      const fileName = path.basename(collectionsFile);
-      // LEGACY, to be removed
-      if (fileName.startsWith("mocks")) {
-        this._deprecationAlerts.set(
-          "mocks",
-          `Defining collections in '${fileName}' file is deprecated. Please rename it to '${fileName.replace(
-            "mocks",
-            "collections"
-          )}'`
-        );
-      } else {
-        this._deprecationAlerts.remove("mocks");
+          .catch((error) => {
+            this._alertsLoad.set(filePath, `Error loading file ${filePath}`, error);
+            return Promise.resolve({
+              path: filePath,
+              error,
+            });
+          });
+      })
+    ).then((filesDetails) => {
+      const loadedFiles = filesDetails.filter((fileDetails) => !!fileDetails.content);
+      const erroredFiles = filesDetails.filter((fileDetails) => !!fileDetails.error);
+      const loadProcess = loader.load(loadedFiles, erroredFiles, {
+        alerts: loader.alerts,
+        logger: loader.logger,
+      });
+      if (isPromise(loadProcess)) {
+        return loadProcess.catch((error) => {
+          this._alertsLoad.set(loader.id, `Error proccesing loaded files`, error);
+          return Promise.resolve();
+        });
       }
-      try {
-        const collections = this._readFile(collectionsFile);
-        const fileErrors = validateFileContent(collections);
-        if (!!fileErrors) {
-          throw new Error(fileErrors);
-        }
-        this._loadCollections(collections);
-        this._logger.silly(`Loaded collections from file ${collectionsFile}`);
-        this._collectionsAlerts.clean();
-      } catch (error) {
-        this._loadCollections([]);
-        this._collectionsAlerts.set(
-          "error",
-          `Error loading collections from file ${collectionsFile}`,
-          error
-        );
-      }
-    } else {
-      this._loadCollections([]);
-      this._collectionsAlerts.set("not-found", `No collections file was found in ${this._path}`);
-    }
+      return Promise.resolve();
+    });
   }
 
   _switchWatch() {
@@ -262,11 +266,14 @@ class FilesLoaders {
       this._watcher = watch(
         this._path,
         { recursive: true },
-        debounce(() => {
-          this._logger.info("File change detected");
-          this._loadFiles();
-        }),
-        1000
+        debounce(
+          () => {
+            this._logger.info("File change detected");
+            this._loadFiles();
+          },
+          200,
+          { maxWait: 1000 }
+        )
       );
     }
   }
@@ -282,6 +289,22 @@ class FilesLoaders {
 
   _cache() {
     return this._customRequireCache || require.cache;
+  }
+
+  createLoader({ id, src, onLoad }) {
+    this._logger.debug(`Creating files loader '${id}'`);
+    this._loaders[id] = new Loader({
+      id,
+      alerts: this._alertsLoaders.collection(id),
+      logger: this._loggerLoaders.namespace(id),
+      src,
+      onLoad,
+    });
+    return this._loaders[id];
+  }
+
+  get loaders() {
+    return { ...this._loaders };
   }
 }